+++ src/mapping.ts
@@ -1,9 +1,5 @@
 import { crypto, ByteArray, BigInt, log } from "@graphprotocol/graph-ts"
 import {
-  OperatorBonded,
-  OperatorConfirmed,
-} from "../generated/SimplePREApplication/SimplePREApplication"
-import {
   Staked,
   ToppedUp,
   Unstaked,
@@ -273,23 +269,6 @@ export function handleDelegateVotesChanged(event: DelegateVotesChanged): void {
   const difference = event.params.newBalance.minus(event.params.previousBalance)
   daoMetric.stakedTotal = daoMetric.stakedTotal.plus(difference)
   daoMetric.save()
-}
-
-export function handleOperatorBonded(event: OperatorBonded): void {
-  const preOperator = new PREOperator(
-    event.params.stakingProvider.toHexString()
-  )
-  preOperator.operator = event.params.operator
-  preOperator.save()
-}
-
-export function handleOperatorConfirmed(event: OperatorConfirmed): void {
-  const preOperator = new PREOperator(
-    event.params.stakingProvider.toHexString()
-  )
-  preOperator.operator = event.params.operator
-  preOperator.confirmationTimestamp = event.block.timestamp
-  preOperator.save()
 }
 
 export function handleMinStakeAmountChanged(
